requirements.txt
flask
langgraph
mysql-connector-python
psycopg2
transformers
torch
